[FEATURE] Allow build-time defined main entrypoint that can be non-executable
Nothing changes in the current behaviour regarding main file.
If the path to main is derived from `bin` settings of project's
`composer.json` it's treated as executable - same way as before.

There are two new commandline options for the build command:
--main - allows to define relative path to the main script at build-time
--executable - forces the build-time defined main script to be treate as executable

If --main is set but --executable is not then the PHAR will not be made executable:
neither the excutability bit will be added to the target file nor shebang will be
prepended.

Rationale:
  - Allow building multiple PHARs with different entrypoint scripts using the
    same project codebase.
  - Allow building PHARs which can be directly included or prepended into
    an PHP application.

The second use case was not possible since prepending the shebang prevents
inclusion via the standard include/require mechanism.